Adapted to new structure of MSharp folders
